// src/app/model/student-details.ts
export class StudentDetails {
    studentId: String;
    stdFirstName: String;
    stdMiddleName: String;
    stdLastName: String;
    fatherFirstName: String;
    fatherMiddleName: String;
    fatherLastName: String;
    dateOfBirth: String;
    gender: String;
    category: String;
    familyIncomeRange: String;
    adharNo: String;
    stdMobNo: String;
    stdPanNo: String;
    stdEmail: String;
    fatherMobNo: String;
    adressLine1: String;
    adressLine2: String;
    city: String;
    district: String;
    state: String;
    country: String;
    studentImage: String;
}